Installed jest-environment-jsdom. FunctionsÑ moved to utils.js фand exported to ьÑmain.js and main.test.js.
