+++ babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  presets: [['@babel/preset-env', {targets: {node: 'current'}}]],
+  presets: [['@babel/preset-env', { targets: { node: 'current' } }]],
 };
