+++ package.json
@@ -23,11 +23,11 @@
   "devDependencies": {
     "@babel/core": "^7.19.3",
     "@babel/preset-env": "^7.19.3",
-    "@ngneat/falso": "^6.1.0",
     "babel-jest": "^29.1.2",
-    "eslint": "^8.23.1",
+    "eslint": "8.22.0",
     "husky": "^8.0.1",
     "jest": "^29.1.2",
+    "jest-environment-jsdom": "^29.1.2",
     "prettier": "2.7.1"
   }
 }
