+++ src/index.html
@@ -1,17 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="description" content="my test project" />
-  <meta name="keywords" content="HTML, CSS, JS, React" />
-  <meta name="author" content="Max" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Title</title>
-</head>
-<body>
-<p id="text1">Text 1</p>
-<p id="text2">Text 2</p>
-<p id="text3">Text 3</p>
-<script type="module" src="scripts/main.js"></script>
-</body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="description" content="my test project" />
+    <meta name="keywords" content="HTML, CSS, JS, React" />
+    <meta name="author" content="Max" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Title</title>
+  </head>
+  <body>
+    <p id="text1">Text 1</p>
+    <p id="text2">Text 2</p>
+    <p id="text3">Text 3</p>
+    <script type="module" src="scripts/main.js"></script>
+    <script type="module" src="tests/main.test.js"></script>
+    <script type="module" src="scripts/utils.js"></script>
+  </body>
 </html>
